Create RemoteShell.cmd
WinRM administration shell to manage remote systems using the command prompt. Suitable for both HTTP and HTTPS connections.
Features:
1. Supports WinRM remoting via HTTP or HTTPS.
2. Connects directly to the C:\ root drive.
3. Input has been sanitized.
4. Standalone shell with interactive menu.
Assumes you have administrative rights on the remote system. Default connection uses HTTP. For HTTPS connections SSL certificates are required on both the target and client. You may refer to online resources on how to configure that. Works out of the box in a domain environment. If target host is in a workgroup or DMZ, addition to trusted hosts is necessary. Refer to WinRM configuration help online.