Create WindowsDefenderGetResults.cmd
This batch script uses the built-in Windows remote shell (winrs.exe) to retrieve the antimalware scan results from the MpCmdRun.log file on a remote computer. WinRM must be enabled on the remote computer, and if it is not part of a domain, its DNS name will need to be added to the trusted hosts list on the local machine. It does not need 3rd party dependencies like PsExec.exe. The MpCmdRun.log file is stored by default on the user's %TEMP% profile.