Create SCEPFullScan.cmd
This batch script uses the built-in Windows remote shell (winrs.exe) to perform a full System Center Endpoint Protection scan on a remote computer. WinRM must be enabled on the remote computer, and if it is not part of a domain, its DNS name will need to be added to the trusted hosts list on the local machine.